backend/requirements.txt
fastapi
uvicorn
livekit-agents[inference]
livekit-api
anthropic
python-dotenv
